Partially fixed poly.line=TRUE & var.col.by==group bug
Fixed a bug with poly.line whereby all lines would be the same linetype
if var.col.by had the same value as group. The issue had to do with the
code setting poly.line.sty based on the value of var.sty.by regardless
of the grouping variable. Introduced a bug in the legend where linetypes
display on the wrong lines when var.sty.by!=NULL.